TMP
so basically in egui-nav we need to do ui.allocate_response to get a
response ahead of time so we can check for hover while the widget is
still being constructed.

The titlebar closure will use a painter, a response and a string (the
title). The response will contain the rect used for painting. it will
also be used to check if the user is hovering

Signed-off-by: kernelkind <kernelkind@gmail.com>